+++ src/main/java/org/rubato/rubettes/alteration/Alterator.java
@@ -186,7 +186,7 @@ public class Alterator {
 		//set starting values
 		Denotator currentDenotator = denotators.next();
 		for (int i = 0; i < paths.length; i++) {
-			double currentDouble = ((RElement)currentDenotator.getElement(paths[i]).cast(RRing.ring)).getValue();
+			double currentDouble = ((RElement)currentDenotator.getElement(paths[i]).cast(RRing.ring)).getValue().doubleValue();
 			minima[i] = currentDouble;
 			maxima[i] = currentDouble;
 		}
@@ -194,7 +194,7 @@ public class Alterator {
 		while (denotators.hasNext()) {
 			currentDenotator = denotators.next();
 			for (int i = 0; i < paths.length; i++) {
-				double currentDouble = ((RElement)currentDenotator.getElement(paths[i]).cast(RRing.ring)).getValue();
+				double currentDouble = ((RElement)currentDenotator.getElement(paths[i]).cast(RRing.ring)).getValue().doubleValue();
 				if (currentDouble > maxima[i]) {
 					maxima[i] = currentDouble;
 				}
@@ -259,7 +259,7 @@ public class Alterator {
 	private double[] getCastDoubleValues(Denotator denotator, int[][] paths) throws RubatoException{
 		double[] values = new double[paths.length];
 		for (int i = 0; i < values.length; i++) {
-			values[i] = ((RElement)denotator.getElement(paths[i]).cast(RRing.ring)).getValue();
+			values[i] = ((RElement)denotator.getElement(paths[i]).cast(RRing.ring)).getValue().doubleValue();
 		}
 		return values;
 	}
