+++ src/main/java/org/rubato/rubettes/wallpaper/JWallpaperView.java
@@ -163,8 +163,8 @@ public class JWallpaperView extends JGeometryView {
 								//four new corners for each x/y combination
 								for (int k = 0; k < originalElements.size(); k++) {
 									ModuleElement currentElement = originalElements.get(k);
-									double x = ((RElement)RRing.ring.cast(currentElement.getComponent(xIndex))).getValue();
-									double y = ((RElement)RRing.ring.cast(currentElement.getComponent(yIndex))).getValue();
+									double x = ((RElement)RRing.ring.cast(currentElement.getComponent(xIndex))).getValue().doubleValue();
+									double y = ((RElement)RRing.ring.cast(currentElement.getComponent(yIndex))).getValue().doubleValue();
 									this.addPoint(x, y);
 								}
 							}
@@ -174,7 +174,7 @@ public class JWallpaperView extends JGeometryView {
 							int xIndex = xIndices.get(i);
 							for (int k = 0; k < originalElements.size(); k++) {
 								ModuleElement currentElement = originalElements.get(k);
-								double x = ((RElement)RRing.ring.cast(currentElement.getComponent(xIndex))).getValue();
+								double x = ((RElement)RRing.ring.cast(currentElement.getComponent(xIndex))).getValue().doubleValue();
 								double y = originalYs.get(k).doubleValue();
 								this.addPoint(x, y);
 							}
@@ -185,7 +185,7 @@ public class JWallpaperView extends JGeometryView {
 							for (int k = 0; k < originalElements.size(); k++) {
 								ModuleElement currentElement = originalElements.get(k);
 								double x = originalXs.get(k).doubleValue();
-								double y = ((RElement)RRing.ring.cast(currentElement.getComponent(yIndex))).getValue();
+								double y = ((RElement)RRing.ring.cast(currentElement.getComponent(yIndex))).getValue().doubleValue();
 								this.addPoint(x, y);
 							}
 						}
